Add suggest list filter in single-select mode.
When an option has already been selected in single-select mode there
is no easy way to type and filter the list of suggestions without
clearing the widget value first. This is also not possible if the
mandatorySelection option is set. This commit adds a filter input
box at the top of the suggestion list.