Fix tabindex implementation.
Previously the tabindex implementation required a specific value and if
omitted sets it to 1. It then incremented the value for each tag picked
and various other elements built into the widget. This is problematic
because if HTML elements surrounding the tagger do not specify a
tabindex then hitting tab skips the tagger elements.