Improve behaviour when clicking on tagger.
Clicking on any area of whitespace within the tagger, including the
input field, will now trigger the suggestion list in both single and
multi select mode. In single-select mode, clicking on the selected tag
will also trigger the suggestion list.

When clearing the selected tag in single-select mode via the (x), the
suggestions list remain open if it is currently open, and the filter
value (if there is one) will be moved to the main input so that
filtering can continue seamlessly.