Display list on click only with single value.
For performance reasons, only display the suggestions list if we click
on the tagger and it has a single value selected (in single value mode).
Functionality to display the suggestion menu on click may be restored
later for multi select or single-select with no value when performance
of generating the suggestion list has been improved.